update usage example
