+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* update usage example
+
 ## 0.0.1
 
 * initial release with iOS support
