+++ example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_native_text_input","path":"/Users/henryleung/flutter-native-text-input/","dependencies":[]}],"android":[],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"flutter_native_text_input","dependencies":[]}],"date_created":"2020-08-07 19:35:43.248184","version":"1.21.0-1.0.pre"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_native_text_input","path":"/Users/henryleung/flutter-native-text-input/","dependencies":[]}],"android":[],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"flutter_native_text_input","dependencies":[]}],"date_created":"2020-08-07 19:59:23.390207","version":"1.21.0-1.0.pre"}
